backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.5
python-socketio>=5.10.0
pylint>=3.0.3
bandit>=1.7.5
autopep8>=2.0.4
javalang>=0.13.0
lxml>=4.9.3
sqlparse>=0.4.4
jsonschema>=4.20.0
pyyaml>=6.0.1
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.0.0
google-generativeai>=0.3.0
anthropic>=0.7.0
tiktoken>=0.5.0
httpx>=0.25.0

